[Fixes #689] Portlets with parameters
Portlets that have parameters (other than id) will now work when being viewed in the edit mode. There may be issues if portlets want an :id paremeter since there is assumed to be the page id. Hopefully this won't be a problem in practice since Addressable content could be used (or upgrading sites can scope the parameter to something thing_id)

* Set up scenarios to cover portlets that have/need parameters
* Add the existing params to the content iframe link.